chore(gamepad): delete dead legacy GilrsProvider
The legacy GilrsProvider in provider.rs was unused in production (zero
references outside the file itself; the production path is
HybridGamepadProvider). It contained a tokio::mpsc::unbounded_channel
plus an untracked std::thread::spawn and tokio::spawn -- landmines if
the file were accidentally rewired into the active code path.

Move the still-used GamepadEvent enum into hybrid_provider/mod.rs
(its natural home, since the hybrid provider is the only producer
in production) and update all imports. Delete provider.rs and remove
its module declaration from gamepad/mod.rs.
